requirements.txt
Flask
Flask-Cors
gevent
scrapy 
beautifulsoup4 
nltk
sumy
numpy
python-dotenv